refactor:서버 로직 분리 1차
