+++ frontend/src/server/constants/packs/a1.ts
@@ -1,0 +1,600 @@
+const CHARIZARD_R1: CardId[] = [
+  'A1-011',
+  'A1-018',
+  'A1-021',
+  'A1-024',
+  'A1-027',
+  'A1-029',
+  'A1-031',
+  'A1-032',
+  'A1-033',
+  'A1-037',
+  'A1-042',
+  'A1-044',
+  'A1-048',
+  'A1-051',
+  'A1-057',
+  'A1-059',
+  'A1-074',
+  'A1-085',
+  'A1-087',
+  'A1-090',
+  'A1-092',
+  'A1-105',
+  'A1-110',
+  'A1-111',
+  'A1-115',
+  'A1-118',
+  'A1-133',
+  'A1-134',
+  'A1-135',
+  'A1-137',
+  'A1-141',
+  'A1-143',
+  'A1-155',
+  'A1-162',
+  'A1-164',
+  'A1-178',
+  'A1-179',
+  'A1-181',
+  'A1-189',
+  'A1-190',
+  'A1-191',
+  'A1-192',
+  'A1-196',
+  'A1-198',
+  'A1-199',
+  'A1-206',
+  'A1-212',
+  'A1-214',
+  'A1-215',
+  'A1-217',
+];
+const PIKACHU_R1: CardId[] = [
+  'A1-005',
+  'A1-006',
+  'A1-014',
+  'A1-027',
+  'A1-029',
+  'A1-039',
+  'A1-042',
+  'A1-048',
+  'A1-051',
+  'A1-053',
+  'A1-057',
+  'A1-064',
+  'A1-070',
+  'A1-072',
+  'A1-073',
+  'A1-077',
+  'A1-092',
+  'A1-094',
+  'A1-097',
+  'A1-099',
+  'A1-101',
+  'A1-105',
+  'A1-110',
+  'A1-111',
+  'A1-113',
+  'A1-118',
+  'A1-124',
+  'A1-133',
+  'A1-134',
+  'A1-135',
+  'A1-137',
+  'A1-139',
+  'A1-147',
+  'A1-160',
+  'A1-162',
+  'A1-164',
+  'A1-166',
+  'A1-169',
+  'A1-179',
+  'A1-189',
+  'A1-190',
+  'A1-193',
+  'A1-194',
+  'A1-198',
+  'A1-199',
+  'A1-208',
+  'A1-212',
+  'A1-214',
+  'A1-215',
+  'A1-216',
+];
+const MEWTWO_R1: CardId[] = [
+  'A1-001',
+  'A1-008',
+  'A1-009',
+  'A1-016',
+  'A1-025',
+  'A1-027',
+  'A1-029',
+  'A1-042',
+  'A1-048',
+  'A1-049',
+  'A1-050',
+  'A1-051',
+  'A1-057',
+  'A1-062',
+  'A1-066',
+  'A1-068',
+  'A1-092',
+  'A1-105',
+  'A1-107',
+  'A1-110',
+  'A1-111',
+  'A1-118',
+  'A1-120',
+  'A1-127',
+  'A1-130',
+  'A1-133',
+  'A1-134',
+  'A1-135',
+  'A1-137',
+  'A1-151',
+  'A1-154',
+  'A1-156',
+  'A1-162',
+  'A1-164',
+  'A1-172',
+  'A1-174',
+  'A1-176',
+  'A1-179',
+  'A1-183',
+  'A1-186',
+  'A1-187',
+  'A1-189',
+  'A1-190',
+  'A1-198',
+  'A1-199',
+  'A1-207',
+  'A1-212',
+  'A1-214',
+  'A1-215',
+  'A1-218',
+];
+
+const CHARIZARD_R2: CardId[] = [
+  'A1-012',
+  'A1-019',
+  'A1-026',
+  'A1-028',
+  'A1-030',
+  'A1-034',
+  'A1-038',
+  'A1-043',
+  'A1-052',
+  'A1-058',
+  'A1-060',
+  'A1-075',
+  'A1-086',
+  'A1-088',
+  'A1-091',
+  'A1-093',
+  'A1-106',
+  'A1-112',
+  'A1-116',
+  'A1-119',
+  'A1-136',
+  'A1-138',
+  'A1-142',
+  'A1-144',
+  'A1-158',
+  'A1-163',
+  'A1-165',
+  'A1-180',
+  'A1-197',
+  'A1-200',
+  'A1-204',
+  'A1-213',
+  'A1-219',
+  'A1-221',
+  'A1-225',
+];
+
+const PIKACHU_R2: CardId[] = [
+  'A1-015',
+  'A1-026',
+  'A1-028',
+  'A1-030',
+  'A1-043',
+  'A1-052',
+  'A1-054',
+  'A1-058',
+  'A1-065',
+  'A1-071',
+  'A1-081',
+  'A1-091',
+  'A1-093',
+  'A1-100',
+  'A1-106',
+  'A1-112',
+  'A1-114',
+  'A1-119',
+  'A1-136',
+  'A1-138',
+  'A1-140',
+  'A1-148',
+  'A1-150',
+  'A1-161',
+  'A1-163',
+  'A1-165',
+  'A1-167',
+  'A1-170',
+  'A1-180',
+  'A1-200',
+  'A1-202',
+  'A1-213',
+  'A1-220',
+  'A1-224',
+  'A1-226',
+];
+const MEWTWO_R2: CardId[] = [
+  'A1-002',
+  'A1-017',
+  'A1-026',
+  'A1-028',
+  'A1-030',
+  'A1-043',
+  'A1-052',
+  'A1-058',
+  'A1-063',
+  'A1-067',
+  'A1-069',
+  'A1-091',
+  'A1-093',
+  'A1-106',
+  'A1-108',
+  'A1-112',
+  'A1-119',
+  'A1-121',
+  'A1-126',
+  'A1-131',
+  'A1-136',
+  'A1-138',
+  'A1-152',
+  'A1-157',
+  'A1-163',
+  'A1-165',
+  'A1-173',
+  'A1-180',
+  'A1-184',
+  'A1-200',
+  'A1-201',
+  'A1-209',
+  'A1-213',
+  'A1-222',
+  'A1-223',
+];
+
+const CHARIZARD_R3: CardId[] = [
+  'A1-013',
+  'A1-020',
+  'A1-022',
+  'A1-035',
+  'A1-045',
+  'A1-046',
+  'A1-061',
+  'A1-079',
+  'A1-089',
+  'A1-117',
+  'A1-145',
+  'A1-159',
+  'A1-182',
+  'A1-203',
+];
+const PIKACHU_R3: CardId[] = [
+  'A1-007',
+  'A1-040',
+  'A1-055',
+  'A1-078',
+  'A1-082',
+  'A1-095',
+  'A1-098',
+  'A1-102',
+  'A1-103',
+  'A1-125',
+  'A1-149',
+  'A1-168',
+  'A1-171',
+  'A1-211',
+];
+const MEWTWO_R3: CardId[] = [
+  'A1-003',
+  'A1-010',
+  'A1-080',
+  'A1-083',
+  'A1-109',
+  'A1-122',
+  'A1-128',
+  'A1-132',
+  'A1-175',
+  'A1-177',
+  'A1-185',
+  'A1-188',
+  'A1-205',
+  'A1-210',
+];
+
+const CHARIZARD_R4: CardId[] = [
+  'A1-023',
+  'A1-036',
+  'A1-047',
+  'A1-076',
+  'A1-146',
+];
+const PIKACHU_R4: CardId[] = ['A1-041', 'A1-056', 'A1-096', 'A1-104', 'A1-195'];
+const MEWTWO_R4: CardId[] = ['A1-004', 'A1-084', 'A1-123', 'A1-129', 'A1-153'];
+
+const CHARIZARD_S1: CardId[] = [
+  'A1-228',
+  'A1-229',
+  'A1-230',
+  'A1-231',
+  'A1-234',
+  'A1-236',
+  'A1-237',
+  'A1-246',
+];
+const PIKACHU_S1: CardId[] = [
+  'A1-232',
+  'A1-233',
+  'A1-235',
+  'A1-238',
+  'A1-240',
+  'A1-241',
+  'A1-248',
+  'A1-250',
+];
+const MEWTWO_S1: CardId[] = [
+  'A1-227',
+  'A1-239',
+  'A1-242',
+  'A1-243',
+  'A1-244',
+  'A1-245',
+  'A1-247',
+  'A1-249',
+];
+
+const CHARIZARD_S2: CardId[] = [
+  'A1-252',
+  'A1-253',
+  'A1-255',
+  'A1-257',
+  'A1-263',
+  'A1-266',
+  'A1-268',
+  'A1-272',
+  'A1-274',
+  'A1-278',
+];
+const PIKACHU_S2: CardId[] = [
+  'A1-254',
+  'A1-256',
+  'A1-259',
+  'A1-260',
+  'A1-265',
+  'A1-267',
+  'A1-271',
+  'A1-273',
+  'A1-276',
+  'A1-279',
+];
+const MEWTWO_S2: CardId[] = [
+  'A1-251',
+  'A1-258',
+  'A1-261',
+  'A1-262',
+  'A1-264',
+  'A1-269',
+  'A1-270',
+  'A1-275',
+  'A1-277',
+];
+
+const CHARIZARD_S3: CardId[] = ['A1-280'];
+const PIKACHU_S3: CardId[] = ['A1-281'];
+const MEWTWO_S3: CardId[] = ['A1-282'];
+
+const ALL_CROWN: CardId[] = ['A1-284', 'A1-285', 'A1-286'];
+
+export const A1_CHARIZARD_PACK_ID_SET = new Set(
+  ...CHARIZARD_R1,
+  ...CHARIZARD_R2,
+  ...CHARIZARD_R3,
+  ...CHARIZARD_R4,
+  ...CHARIZARD_S1,
+  ...CHARIZARD_S2,
+  ...CHARIZARD_S3,
+  ALL_CROWN
+);
+export const A1_PIKACHU_PACK_ID_SET = new Set(
+  ...PIKACHU_R1,
+  ...PIKACHU_R2,
+  ...PIKACHU_R3,
+  ...PIKACHU_R4,
+  ...PIKACHU_S1,
+  ...PIKACHU_S2,
+  ...PIKACHU_S3,
+  ALL_CROWN
+);
+export const A1_MEWTWO_PACK_ID_SET = new Set(
+  ...MEWTWO_R1,
+  ...MEWTWO_R2,
+  ...MEWTWO_R3,
+  ...MEWTWO_R4,
+  ...MEWTWO_S1,
+  ...MEWTWO_S2,
+  ...MEWTWO_S3,
+  ALL_CROWN
+);
+
+export const PIKACHU_PACK_CARD_ID_LIST: Record<Rarity, CardId[]>[] = [
+  {
+    crown: [],
+    s3: [],
+    s2: [],
+    s1: [],
+    r4: [],
+    r3: [],
+    r2: [],
+    r1: PIKACHU_R1,
+  },
+  {
+    crown: [],
+    s3: [],
+    s2: [],
+    s1: [],
+    r4: [],
+    r3: [],
+    r2: [],
+    r1: PIKACHU_R1,
+  },
+  {
+    crown: [],
+    s3: [],
+    s2: [],
+    s1: [],
+    r4: [],
+    r3: [],
+    r2: [],
+    r1: PIKACHU_R1,
+  },
+  {
+    crown: ALL_CROWN,
+    s3: PIKACHU_S3,
+    s2: PIKACHU_S2,
+    s1: PIKACHU_S1,
+    r4: PIKACHU_R4,
+    r3: PIKACHU_R3,
+    r2: PIKACHU_R2,
+    r1: [],
+  },
+  {
+    crown: ALL_CROWN,
+    s3: PIKACHU_S3,
+    s2: PIKACHU_S2,
+    s1: PIKACHU_S1,
+    r4: PIKACHU_R4,
+    r3: PIKACHU_R3,
+    r2: PIKACHU_R2,
+    r1: [],
+  },
+];
+
+export const MEWTWO_PACK_CARD_ID_LIST: Record<Rarity, CardId[]>[] = [
+  {
+    crown: [],
+    s3: [],
+    s2: [],
+    s1: [],
+    r4: [],
+    r3: [],
+    r2: [],
+    r1: MEWTWO_R1,
+  },
+  {
+    crown: [],
+    s3: [],
+    s2: [],
+    s1: [],
+    r4: [],
+    r3: [],
+    r2: [],
+    r1: MEWTWO_R1,
+  },
+  {
+    crown: [],
+    s3: [],
+    s2: [],
+    s1: [],
+    r4: [],
+    r3: [],
+    r2: [],
+    r1: MEWTWO_R1,
+  },
+  {
+    crown: ALL_CROWN,
+    s3: MEWTWO_S3,
+    s2: MEWTWO_S2,
+    s1: MEWTWO_S1,
+    r4: MEWTWO_R4,
+    r3: MEWTWO_R3,
+    r2: MEWTWO_R2,
+    r1: [],
+  },
+  {
+    crown: ALL_CROWN,
+    s3: MEWTWO_S3,
+    s2: MEWTWO_S2,
+    s1: MEWTWO_S1,
+    r4: MEWTWO_R4,
+    r3: MEWTWO_R3,
+    r2: MEWTWO_R2,
+    r1: [],
+  },
+];
+
+export const CHARIZARD_PACK_CARD_ID_LIST: Record<Rarity, CardId[]>[] = [
+  {
+    crown: [],
+    s3: [],
+    s2: [],
+    s1: [],
+    r4: [],
+    r3: [],
+    r2: [],
+    r1: CHARIZARD_R1,
+  },
+  {
+    crown: [],
+    s3: [],
+    s2: [],
+    s1: [],
+    r4: [],
+    r3: [],
+    r2: [],
+    r1: CHARIZARD_R1,
+  },
+  {
+    crown: [],
+    s3: [],
+    s2: [],
+    s1: [],
+    r4: [],
+    r3: [],
+    r2: [],
+    r1: CHARIZARD_R1,
+  },
+  {
+    crown: ALL_CROWN,
+    s3: CHARIZARD_S3,
+    s2: CHARIZARD_S2,
+    s1: CHARIZARD_S1,
+    r4: CHARIZARD_R4,
+    r3: CHARIZARD_R3,
+    r2: CHARIZARD_R2,
+    r1: [],
+  },
+  {
+    crown: ALL_CROWN,
+    s3: CHARIZARD_S3,
+    s2: CHARIZARD_S2,
+    s1: CHARIZARD_S1,
+    r4: CHARIZARD_R4,
+    r3: CHARIZARD_R3,
+    r2: CHARIZARD_R2,
+    r1: [],
+  },
+];
+
+export const A1_CARD_POOL_ID_LIST: Record<
+  A1PackType,
+  Record<Rarity, CardId[]>[]
+> = {
+  charizard: CHARIZARD_PACK_CARD_ID_LIST,
+  pikachu: PIKACHU_PACK_CARD_ID_LIST,
+  mewtwo: MEWTWO_PACK_CARD_ID_LIST,
+};
