+++ frontend/src/types/logicType.d.ts
@@ -1,0 +1,3 @@
+type PercentTuple<T> = [T, number][];
+
+type RarityPercentTuple = PercentTuple<Rarity>;
